split river/lake/ocean-related functions from overmap.cpp
